model change

+++ zombie_game_type/model/resource_drop_probability.py
@@ -8,4 +8,4 @@ from ..base_model import BaseModel
 class ResourceDropProbabilityModel(BaseModel):
     name: str
     resource: ResourceModel
-    probability: int
+    probability: float
